// actions/src/lib/github/api/repos/response/topic.ts
/* tslint:disable */
/**
 * This file was automatically generated by json-schema-to-typescript.
 * DO NOT MODIFY IT BY HAND. Instead, modify the source JSONSchema file,
 * and run json-schema-to-typescript to regenerate this file.
 */

/**
 * A topic aggregates entities that are related to a subject.
 */
export interface Topic {
  names: string[];
  [k: string]: unknown;
}